commit- se realiza commit y ajustes al pipline

+++ Jenkinsfile
@@ -82,6 +82,7 @@ pipeline {
 
         stage('Run Tests') {
             parallel {
+                
                 stage('Test Frontend') {
                     steps {
                         dir('./proyecto-frontApp') {
@@ -98,6 +99,7 @@ pipeline {
                         }
                     }
                 }
+
                 stage('Test BFF') {
                     steps {
                         dir('./ms-nestjs-bff') {
@@ -110,6 +112,7 @@ pipeline {
                         }
                     }
                 }
+
                 stage('Test Security') {
                     steps {
                         dir('./ms-nestjs-security') {
@@ -122,19 +125,24 @@ pipeline {
                         }
                     }
                 }
+
                 stage('Test Python') {
                     steps {
                         dir('./ms-python') {
-                            script {
-                                echo "Configurando entorno virtual de Python..."
-                                sh 'python3 -m venv venv && . venv/bin/activate'
-
-                                echo "Instalando dependencias de Python..."
-                                sh 'pip install -r requirements.txt'
-
-                                echo "Ejecutando pruebas de Python..."
-                                sh 'pytest'
-                            }
+                                script {
+                                    echo "Configurando entorno virtual de Python..."
+                                    sh '''
+                                        if [ -f venv/bin/activate ]; then
+                                            . venv/bin/activate
+                                            elif [ -f venv\\Scripts\\activate ]; then
+                                            . venv\\Scripts\\activate
+                                            fi
+                                            '''
+                                            echo "Instalando dependencias de Python..."
+                                            sh 'pip install -r requirements.txt'
+                                            echo "Ejecutando pruebas de Python..."
+                                            sh 'pytest > result.log || tail -n 10 result.log'
+                                }
                         }
                     }
                 }
